removing visualize_data_loader and request/response handlers (#44769)
