+++ src/legacy/core_plugins/interpreter/index.ts
@@ -30,13 +30,11 @@ export default function InterpreterPlugin(kibana: any) {
     uiExports: {
       injectDefaultVars: server => ({
         serverBasePath: server.config().get('server.basePath'),
-        interpreterConfig: server.config().get('interpreter'),
       }),
     },
     config: (Joi: any) => {
       return Joi.object({
         enabled: Joi.boolean().default(true),
-        enableInVisualize: Joi.boolean().default(true),
       }).default();
     },
     init,
