+++ src/legacy/core_plugins/interpreter/public/functions/esaggs.ts
@@ -17,9 +17,8 @@
  * under the License.
  */
 
-import { get } from 'lodash';
+import { get, has } from 'lodash';
 import { i18n } from '@kbn/i18n';
-import { CourierRequestHandlerProvider } from 'ui/vis/request_handlers/courier';
 // @ts-ignore
 import { AggConfigs } from 'ui/vis/agg_configs';
 import { createFormat } from 'ui/visualize/loader/pipeline_helpers/utilities';
@@ -27,15 +26,25 @@ import chrome from 'ui/chrome';
 
 // need to get rid of angular from these
 // @ts-ignore
-import { SearchSourceProvider } from 'ui/courier/search_source';
-import { FilterBarQueryFilterProvider } from 'ui/filter_manager/query_filter';
-import { IndexPatternsProvider } from '../../../data/public';
+import { TimeRange } from 'src/plugins/data/public';
+import { SearchSource } from '../../../../ui/public/courier/search_source';
+// @ts-ignore
+import { SearchSourceProvider } from '../../../../ui/public/courier/search_source';
+import { FilterBarQueryFilterProvider } from '../../../../ui/public/filter_manager/query_filter';
 
-const courierRequestHandlerProvider = CourierRequestHandlerProvider;
-const courierRequestHandler = courierRequestHandlerProvider().handler;
-
-import { ExpressionFunction, KibanaDatatableColumn } from '../../types';
+import { buildTabularInspectorData } from '../../../../ui/public/inspector/build_tabular_inspector_data';
+import {
+  getRequestInspectorStats,
+  getResponseInspectorStats,
+} from '../../../../ui/public/courier/utils/courier_inspector_utils';
+import { calculateObjectHash } from '../../../../ui/public/vis/lib/calculate_object_hash';
+import { getTime } from '../../../../ui/public/timefilter';
+import { RequestHandlerParams } from '../../../../ui/public/visualize/loader/embedded_visualize_handler';
+// @ts-ignore
+import { tabifyAggResponse } from '../../../../ui/public/agg_response/tabify/tabify';
 import { KibanaContext, KibanaDatatable } from '../../common';
+import { ExpressionFunction, KibanaDatatableColumn } from '../../types';
+import { IndexPatternsProvider } from '../../../data/public';
 
 const name = 'esaggs';
 
@@ -50,6 +59,156 @@ interface Arguments {
 }
 
 type Return = Promise<KibanaDatatable>;
+
+const handleCourierRequest = async ({
+  searchSource,
+  aggs,
+  timeRange,
+  query,
+  filters,
+  forceFetch,
+  partialRows,
+  metricsAtAllLevels,
+  inspectorAdapters,
+  queryFilter,
+  abortSignal,
+}: RequestHandlerParams) => {
+  // Create a new search source that inherits the original search source
+  // but has the appropriate timeRange applied via a filter.
+  // This is a temporary solution until we properly pass down all required
+  // information for the request to the request handler (https://github.com/elastic/kibana/issues/16641).
+  // Using callParentStartHandlers: true we make sure, that the parent searchSource
+  // onSearchRequestStart will be called properly even though we use an inherited
+  // search source.
+  const timeFilterSearchSource = searchSource.createChild({ callParentStartHandlers: true });
+  const requestSearchSource = timeFilterSearchSource.createChild({ callParentStartHandlers: true });
+
+  aggs.setTimeRange(timeRange as TimeRange);
+
+  // For now we need to mirror the history of the passed search source, since
+  // the request inspector wouldn't work otherwise.
+  Object.defineProperty(requestSearchSource, 'history', {
+    get() {
+      return searchSource.history;
+    },
+    set(history) {
+      return (searchSource.history = history);
+    },
+  });
+
+  requestSearchSource.setField('aggs', function() {
+    return aggs.toDsl(metricsAtAllLevels);
+  });
+
+  requestSearchSource.onRequestStart((paramSearchSource: SearchSource, searchRequest: unknown) => {
+    return aggs.onSearchRequestStart(paramSearchSource, searchRequest);
+  });
+
+  if (timeRange) {
+    timeFilterSearchSource.setField('filter', () => {
+      return getTime(searchSource.getField('index'), timeRange);
+    });
+  }
+
+  requestSearchSource.setField('filter', filters);
+  requestSearchSource.setField('query', query);
+
+  const reqBody = await requestSearchSource.getSearchRequestBody();
+
+  const queryHash = calculateObjectHash(reqBody);
+  // We only need to reexecute the query, if forceFetch was true or the hash of the request body has changed
+  // since the last request
+  const shouldQuery = forceFetch || searchSource.lastQuery !== queryHash;
+
+  if (shouldQuery) {
+    inspectorAdapters.requests.reset();
+    const request = inspectorAdapters.requests.start(
+      i18n.translate('interpreter.functions.esaggs.inspector.dataRequest.title', {
+        defaultMessage: 'Data',
+      }),
+      {
+        description: i18n.translate(
+          'interpreter.functions.esaggs.inspector.dataRequest.description',
+          {
+            defaultMessage:
+              'This request queries Elasticsearch to fetch the data for the visualization.',
+          }
+        ),
+      }
+    );
+    request.stats(getRequestInspectorStats(requestSearchSource));
+
+    try {
+      // Abort any in-progress requests before fetching again
+      if (abortSignal) {
+        abortSignal.addEventListener('abort', () => requestSearchSource.cancelQueued());
+      }
+
+      const response = await requestSearchSource.fetch();
+
+      searchSource.lastQuery = queryHash;
+
+      request.stats(getResponseInspectorStats(searchSource, response)).ok({ json: response });
+
+      searchSource.rawResponse = response;
+    } catch (e) {
+      // Log any error during request to the inspector
+      request.error({ json: e });
+      throw e;
+    } finally {
+      // Add the request body no matter if things went fine or not
+      requestSearchSource.getSearchRequestBody().then((req: unknown) => {
+        request.json(req);
+      });
+    }
+  }
+
+  // Note that rawResponse is not deeply cloned here, so downstream applications using courier
+  // must take care not to mutate it, or it could have unintended side effects, e.g. displaying
+  // response data incorrectly in the inspector.
+  let resp = searchSource.rawResponse;
+  for (const agg of aggs.aggs) {
+    if (has(agg, 'type.postFlightRequest')) {
+      resp = await agg.type.postFlightRequest(
+        resp,
+        aggs,
+        agg,
+        requestSearchSource,
+        inspectorAdapters,
+        abortSignal
+      );
+    }
+  }
+
+  searchSource.finalResponse = resp;
+
+  const parsedTimeRange = timeRange ? getTime(aggs.indexPattern, timeRange) : null;
+  const tabifyParams = {
+    metricsAtAllLevels,
+    partialRows,
+    timeRange: parsedTimeRange ? parsedTimeRange.range : undefined,
+  };
+
+  const tabifyCacheHash = calculateObjectHash({ tabifyAggs: aggs, ...tabifyParams });
+  // We only need to reexecute tabify, if either we did a new request or some input params to tabify changed
+  const shouldCalculateNewTabify = shouldQuery || searchSource.lastTabifyHash !== tabifyCacheHash;
+
+  if (shouldCalculateNewTabify) {
+    searchSource.lastTabifyHash = tabifyCacheHash;
+    searchSource.tabifiedResponse = tabifyAggResponse(
+      aggs,
+      searchSource.finalResponse,
+      tabifyParams
+    );
+  }
+
+  inspectorAdapters.data.setTabularLoader(
+    () => buildTabularInspectorData(searchSource.tabifiedResponse, queryFilter),
+    { returnsFormattedValues: true }
+  );
+
+  return searchSource.tabifiedResponse;
+};
 
 export const esaggs = (): ExpressionFunction<typeof name, Context, Arguments, Return> => ({
   name,
@@ -91,7 +250,7 @@ export const esaggs = (): ExpressionFunction<typeof name, Context, Arguments, Re
     const $injector = await chrome.dangerouslyGetActiveInjector();
     const Private: Function = $injector.get('Private');
     const indexPatterns = Private(IndexPatternsProvider);
-    const SearchSource = Private(SearchSourceProvider);
+    const SearchSourceClass = Private(SearchSourceProvider);
     const queryFilter = Private(FilterBarQueryFilterProvider);
 
     const aggConfigsState = JSON.parse(args.aggConfigs);
@@ -99,11 +258,11 @@ export const esaggs = (): ExpressionFunction<typeof name, Context, Arguments, Re
     const aggs = new AggConfigs(indexPattern, aggConfigsState);
 
     // we should move searchSource creation inside courier request handler
-    const searchSource = new SearchSource();
+    const searchSource = new SearchSourceClass();
     searchSource.setField('index', indexPattern);
     searchSource.setField('size', 0);
 
-    const response = await courierRequestHandler({
+    const response = await handleCourierRequest({
       searchSource,
       aggs,
       timeRange: get(context, 'timeRange', undefined),
