+++ src/legacy/core_plugins/interpreter/public/functions/visualization.js
@@ -20,26 +20,10 @@
 import { get } from 'lodash';
 import { i18n } from '@kbn/i18n';
 import chrome from 'ui/chrome';
-import { VisRequestHandlersRegistryProvider as RequestHandlersProvider } from 'ui/registry/vis_request_handlers';
-import { VisResponseHandlersRegistryProvider as ResponseHandlerProvider } from 'ui/registry/vis_response_handlers';
 import { VisTypesRegistryProvider } from 'ui/registry/vis_types';
 import { IndexPatternsProvider } from '../../../data/public';
 import { FilterBarQueryFilterProvider } from 'ui/filter_manager/query_filter';
 import { PersistedState } from 'ui/persisted_state';
-
-function getHandler(from, type) {
-  if (typeof type === 'function') {
-    return type;
-  }
-  if (type === 'courier' || type === 'none') {
-    return null;
-  }
-  const handlerDesc = from.find(handler => handler.name === type);
-  if (!handlerDesc) {
-    throw new Error(`Could not find handler "${type}".`);
-  }
-  return handlerDesc.handler;
-}
 
 export const visualization = () => ({
   name: 'visualization',
@@ -80,8 +64,6 @@ export const visualization = () => ({
   async fn(context, args, handlers) {
     const $injector = await chrome.dangerouslyGetActiveInjector();
     const Private = $injector.get('Private');
-    const requestHandlers = Private(RequestHandlersProvider);
-    const responseHandlers = Private(ResponseHandlerProvider);
     const visTypes = Private(VisTypesRegistryProvider);
     const indexPatterns = Private(IndexPatternsProvider);
     const queryFilter = Private(FilterBarQueryFilterProvider);
@@ -89,15 +71,13 @@ export const visualization = () => ({
     const visConfigParams = JSON.parse(args.visConfig);
     const schemas = JSON.parse(args.schemas);
     const visType = visTypes.byName[args.type || 'histogram'];
-    const requestHandler = getHandler(requestHandlers, visType.requestHandler);
-    const responseHandler = getHandler(responseHandlers, visType.responseHandler);
     const indexPattern = args.index ? await indexPatterns.get(args.index) : null;
 
     const uiStateParams = JSON.parse(args.uiState);
     const uiState = new PersistedState(uiStateParams);
 
-    if (requestHandler) {
-      context = await requestHandler({
+    if (typeof visType.requestHandler === 'function') {
+      context = await visType.requestHandler({
         partialRows: args.partialRows,
         metricsAtAllLevels: args.metricsAtAllLevels,
         index: indexPattern,
@@ -112,7 +92,7 @@ export const visualization = () => ({
       });
     }
 
-    if (responseHandler) {
+    if (typeof visType.responseHandler === 'function') {
       if (context.columns) {
         // assign schemas to aggConfigs
         context.columns.forEach(column => {
@@ -130,7 +110,7 @@ export const visualization = () => ({
         });
       }
 
-      context = await responseHandler(context, visConfigParams.dimensions);
+      context = await visType.responseHandler(context, visConfigParams.dimensions);
     }
 
     return {
