+++ src/legacy/core_plugins/kibana/public/kibana.js
@@ -27,8 +27,7 @@ import { uiModules } from 'ui/modules';
 // import the uiExports that we want to "use"
 import 'uiExports/home';
 import 'uiExports/visTypes';
-import 'uiExports/visResponseHandlers';
-import 'uiExports/visRequestHandlers';
+
 import 'uiExports/visEditorTypes';
 import 'uiExports/visualize';
 import 'uiExports/savedObjectTypes';
