+++ src/legacy/core_plugins/kibana/public/visualize/embeddable/visualize_embeddable_factory.tsx
@@ -33,8 +33,6 @@ import 'uiExports/savedObjectTypes';
 import 'uiExports/search';
 import 'uiExports/shareContextMenuExtensions';
 import 'uiExports/visEditorTypes';
-import 'uiExports/visRequestHandlers';
-import 'uiExports/visResponseHandlers';
 import 'uiExports/visTypes';
 import 'uiExports/visualize';
