+++ src/legacy/core_plugins/tests_bundle/tests_entry_template.js
@@ -127,9 +127,6 @@ const coreSystem = new CoreSystem({
         enabled: true,
         enableExternalUrls: true
       },
-      interpreterConfig: {
-        enableInVisualize: true
-      }
     },
   },
   rootDomElement,
