+++ src/legacy/ui/public/vis/index.d.ts
@@ -20,5 +20,3 @@
 export { AggConfig } from './agg_config';
 export { Vis, VisProvider, VisParams, VisState } from './vis';
 export { VisualizationController, VisType } from './vis_types/vis_type';
-export * from './request_handlers';
-export * from './response_handlers';
