--- src/legacy/ui/public/vis/response_handlers/none.js
@@ -1,40 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-// response handler:
-// receives response data and vis configuration
-// returns a promise
-// promise returns response data when resolved
-
-import { VisResponseHandlersRegistryProvider } from '../../registry/vis_response_handlers';
-
-const noneResponseHandler = function () {
-  return {
-    name: 'none',
-    handler: function (response) {
-      return new Promise((resolve) => {
-        resolve(response);
-      });
-    }
-  };
-};
-
-VisResponseHandlersRegistryProvider.register(noneResponseHandler);
-
-export { noneResponseHandler };
