+++ src/legacy/ui/public/visualize/loader/visualize_loader.ts
@@ -30,11 +30,7 @@ import { EmbeddedVisualizeHandler } from './embedded_visualize_handler';
 import { VisSavedObject, VisualizeLoaderParams } from './types';
 
 export class VisualizeLoader {
-  constructor(
-    private readonly savedVisualizations: any,
-    private readonly pipelineDataLoader: boolean,
-    private readonly Private: IPrivate
-  ) {}
+  constructor(private readonly savedVisualizations: any, private readonly Private: IPrivate) {}
 
   /**
    * Renders a saved visualization specified by its id into a DOM element.
@@ -139,19 +135,14 @@ export class VisualizeLoader {
       queryFilter: this.Private(FilterBarQueryFilterProvider),
       // lets add Private to the params, we'll need to pass it to visualize later
       Private: this.Private,
-      pipelineDataLoader: this.pipelineDataLoader,
     };
 
     return new EmbeddedVisualizeHandler(element, savedObj, handlerParams);
   }
 }
 
-function VisualizeLoaderProvider(
-  savedVisualizations: any,
-  interpreterConfig: any,
-  Private: IPrivate
-) {
-  return new VisualizeLoader(savedVisualizations, interpreterConfig.enableInVisualize, Private);
+function VisualizeLoaderProvider(savedVisualizations: any, Private: IPrivate) {
+  return new VisualizeLoader(savedVisualizations, Private);
 }
 
 /**
