+++ src/legacy/ui/ui_exports/ui_export_types/index.js
@@ -41,8 +41,6 @@ export {
 
 export {
   visTypes,
-  visResponseHandlers,
-  visRequestHandlers,
   visEditorTypes,
   interpreter,
   savedObjectTypes,
